Moved View Models to separate classes
They were cluttering up the `TopicMappingServiceTest` class, so moved them to their own standalone files.
