+++ Ignia.Topics.Tests/Ignia.Topics.Tests.csproj
@@ -83,6 +83,10 @@
     <Compile Include="TopicTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="AttributeValueCollectionTest.cs" />
+    <Compile Include="ViewModels\DefaultPropertyTopicViewModel.cs" />
+    <Compile Include="ViewModels\RequiredObjectTopicViewModel.cs" />
+    <Compile Include="ViewModels\RequiredTopicViewModel.cs" />
+    <Compile Include="ViewModels\SampleTopicViewModel.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Ignia.Topics.Data.Caching\Ignia.Topics.Data.Caching.csproj">
