+++ Ignia.Topics.Tests/ViewModels/RequiredTopicViewModel.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Ignia.Topics.Tests.ViewModels {
+
+  public class RequiredTopicViewModel : SampleTopicViewModel {
+
+    [Required]
+    public string RequiredAttribute { get; set; }
+
+  } //Class
+} //Namespace
